Utility for input & output typed test
